Based on remote updates, the function name has been modified and the associated program calls have been ensured to function properly
