+++ TOVsolver/maxium_central_density.py
@@ -2,7 +2,7 @@ import numpy as np
 from scipy.interpolate import interp1d
 
 from TOVsolver import unit
-from TOVsolver.main import OutputMRpoint_with_EoS as MR_point
+from TOVsolver.main import OutputMR
 
 def maxium_central_density(energy_density, pressure, central_densitys=np.logspace(14.3, 15.6, 5) * unit.g_cm_3, num2=30):
     """Outputs the maxium central density of a stable EoS
@@ -14,37 +14,13 @@ def maxium_central_density(energy_density, pressure, central_densitys=np.logspac
     Returns:
         density (float): The maxium central density, in unit.g_cm_3
     """
-    # Notice that we only rescale quantities inside this function
-    energy_density = energy_density * unit.G / unit.c**2
-    pressure = pressure * unit.G / unit.c**4
-
-    # eos = UnivariateSpline(np.log10(energy_density), np.log10(pres), k=1, s=0)
-    # inveos = UnivariateSpline(np.log10(pres), np.log10(energy_density), k=1, s=0)
-    # We could change this to Double Log Interpolation。
-
-    unique_pressure_indices = np.unique(pressure, return_index=True)[1]
-    unique_pressure = pressure[np.sort(unique_pressure_indices)]
-
-    # Interpolate pressure vs. energy density
-    eos = interp1d(energy_density, pressure, kind="cubic", fill_value="extrapolate")
-
-    # Interpolate energy density vs. pressure
-    inveos = interp1d(
-        unique_pressure,
-        energy_density[unique_pressure_indices],
-        kind="cubic",
-        fill_value="extrapolate",
-    )
-
-    Pmin = pressure[20]
-
     ############## Below is the main part of two searches for peaks
     ######## First search
     Ms = [-1,-2] # Meaningless initialization, only to ensure that the following 'if (i>0) and (Ms [-1]<=Ms [-2])' statements can run properly
     store_d_range = [central_densitys[-2], central_densitys[-1]] # When the following loop does not output a result, initialization here will have its meaning
     # Find the extremum point within the predetermined range of central density and return the central density near the extremum point
     for i, rho in enumerate(central_densitys):
-        M, R = MR_point(rho, Pmin, eos, inveos)
+        M, R = OutputMR('', energy_density, pressure, [rho])[0]
         Ms.append(M)
         if (i>0) and (Ms[-1] <= Ms[-2]):
             store_d_range = [central_densitys[i-2], central_densitys[i]] 
@@ -57,7 +33,7 @@ def maxium_central_density(energy_density, pressure, central_densitys=np.logspac
     store_d_range = np.geomspace(store_d_range[0], store_d_range[1], num2) # At least 5 points
     # Note that the first and last points have already been calculated, so there is no need to traverse them
     for i, rho in enumerate(store_d_range[1:-1]):
-        M, R = MR_point(rho, Pmin, eos, inveos)
+        M, R = OutputMR('', energy_density, pressure, [rho])[0]
         Ms.append(M)
         if Ms[-1] <= Ms[-2]:    # Note that due to the second peak search refining the interval, the result is generally not obtained at the first point.
                                 # Therefore, initializing Ms=[-1, -2] is acceptable
