+++ docs/source/main.py
@@ -6,6 +6,7 @@ from scipy.constants import pi
 from scipy.integrate import odeint, ode
 from matplotlib import pyplot
 from scipy import optimize
+from scipy.interpolate import interp1d
 from itertools import repeat
 import csv
 
@@ -14,10 +15,10 @@ import csv
 import TOVsolver.solver_code as TOV_solver
 import TOVsolver.EoS_import as EoS_import
 import TOVsolver.speed_of_sound as speed_of_sound
+from TOVsolver.unit import g_cm_3, G, c
 
 # Global Variables
-def OutputMR(input_file='',density=[],pressure=[]):
-
+def OutputMR(input_file="", density=[], pressure=[], central_density_range=np.logspace(14.3, 15.6, 50) * g_cm_3):
     """Outputs the mass, radius, and tidal deformability
     Args:
         file_name (string, optional): string. CSV file to be opened.
@@ -28,37 +29,55 @@ def OutputMR(input_file='',density=[],pressure=[]):
         MR (tuple): tuple with mass, radius.
     """
 
-    c = 3e10
-    G = 6.67428e-8
-    Msun = 1.989e33
-
-    dyncm2_to_MeVfm3 = 1./(1.6022e33)
-    gcm3_to_MeVfm3 = 1./(1.7827e12)
-    oneoverfm_MeV = 197.33
     #############This is something we need to change, like the input for this EOS import should
     ############# be one file contatining Whole EOS. that first column is density and second is pressure
-    energy_density, pressure = EoS_import.EOS_import(input_file,density,pressure)
+    energy_density, pressure = EoS_import.EOS_import(input_file, density, pressure)
     ############# Lets the user only input the EOS file path, then this EOS_import should have file
     ############# as input. and the outputMR should have a file as input too?
-    
+
+    # Notice that we only rescale quantities inside this function
+    energy_density = energy_density * G / c**2
+    pressure = pressure * G / c**4
+
+    # eos = UnivariateSpline(np.log10(energy_density), np.log10(pres), k=1, s=0)
+    # inveos = UnivariateSpline(np.log10(pres), np.log10(energy_density), k=1, s=0)
+    # We could change this to Double Log Interpolation。
+
+    unique_pressure_indices = np.unique(pressure, return_index=True)[1]
+    unique_pressure = pressure[np.sort(unique_pressure_indices)]
+
+    # Interpolate pressure vs. energy density
+    eos = interp1d(energy_density, pressure, kind="cubic", fill_value="extrapolate")
+
+    # Interpolate energy density vs. pressure
+    inveos = interp1d(
+        unique_pressure,
+        energy_density[unique_pressure_indices],
+        kind="cubic",
+        fill_value="extrapolate",
+    )
+
+    Pmin = pressure[20]
+
     Radius = []
     Mass = []
-    
-    density = np.logspace(14.3, 15.6, 50)
-#This following step is to make a dicision whether the EOS ingredients is always increase. We can do that outsie of this main to the 
-#EOS import.
-#if   all(x<y for x, y in zip(eps_total_poly[:], eps_total_poly[[1:])) and all(x<y for x, y in zip(pres_total_poly[j][:], pres_total_poly[j][1:])):
-    for i in range(len(density)):
+
+    # This following step is to make a dicision whether the EOS ingredients is always increase. We can do that outsie of this main to the
+    # EOS import.
+    # if   all(x<y for x, y in zip(eps_total_poly[:], eps_total_poly[[1:])) and all(x<y for x, y in zip(pres_total_poly[j][:], pres_total_poly[j][1:])):
+    for i in range(len(central_density_range)):
         try:
-            Radius.append(TOV_solver.solveTOV(density[i], energy_density, pressure)[1])
-            Mass.append(TOV_solver.solveTOV(density[i], energy_density, pressure)[0])
-    #This is sentense is for avoiding the outflow of the result, like when solveTOV blow up because of ill EOS, we need to stop
+            center_rho = central_density_range[i] * G / c**2
+            M, R = TOV_solver.solveTOV(center_rho, Pmin, eos, inveos)
+            Mass.append(M)
+            Radius.append(R)
+        # This is sentense is for avoiding the outflow of the result, like when solveTOV blow up because of ill EOS, we need to stop
         except OverflowError as e:
-            #print("This EOS is ill-defined to reach an infinity result, that is not phyiscal, No Mass radius will be generated.")
+            # print("This EOS is ill-defined to reach an infinity result, that is not phyiscal, No Mass radius will be generated.")
             break
-    MR = np.vstack((Radius, Mass)).T
-    #print("Mass Radius file will be generated and stored as  2-d array. The first column is Radius, second one is mass")
-    
+    MR = np.vstack((Mass, Radius)).T
+    # print("Mass Radius file will be generated and stored as  2-d array. The first column is mass, second one is Radius")
+
     return MR
 
 def OutputMRT(input_file='',density=[],pressure=[]):
@@ -124,29 +143,6 @@ def OutputC_s(input_file='',density=[],pressure=[]):
     energy_density, pressure = EoS_import.EOS_import(input_file,density,pressure)
     C_s = speed_of_sound.speed_of_sound_calc(energy_density, pressure)
     return C_s
-
-def OutputMRpoint(central_density,energy_density,pressure):
-
-    """Outputs the mass, radius, and tidal deformability (single point)
-    Args:
-        central_density (float): central density that we want to compute
-        density (array, optional): numpy 1Darray. Density of EoS
-        pressure (array, optional): numpy 1Darray. pressure of EoS
-
-    Returns:
-        MR (tuple): tuple with mass, radius.
-    """
-
-    c = 3e10
-    G = 6.67428e-8
-    Msun = 1.989e33
-
-    dyncm2_to_MeVfm3 = 1./(1.6022e33)
-    gcm3_to_MeVfm3 = 1./(1.7827e12)
-    oneoverfm_MeV = 197.33
-    
-    Radius = []
-    Mass = []
     
 #This following step is to make a dicision whether the EOS ingredients is always increase. We can do that outsie of this main to the 
 #EOS import.
